Hover red, shorten time, delete alert, page name

+++ src/index.js
@@ -5,6 +5,12 @@ import firebase from './lib/firebase.js';
 
 //note: can't use fs (file system) bc react stubs out node core modules
 
+//think about the relationship between people interacting. ppl care more when
+//their actions have consequences on other ppl (or outcomes)
+//anonymity -> lowered stakes -> disengagement
+//look at post-structuralism (like flux, john cage, generated poetry, etc).
+
+
 let dbRef = firebase.firestore().collection('words');
 
 function UpdateWords() { //API call to firebase
@@ -53,8 +59,8 @@ class Master extends React.Component {
       startTime: 0,
       waitLeft: 0
     };
-
-    this.handleChange = this.handleChange.bind(this); //other way of writing is onSubmit={(event) => {this.handleChange(event)}}
+    //other way of writing is onSubmit={(event) => {this.handleChange(event)}}
+    this.handleChange = this.handleChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
   }
 
@@ -62,7 +68,7 @@ class Master extends React.Component {
     //check if startTime exists from prev session
     if (localStorage.getItem('startTime')) {
       //check if 15 min has passed
-      const timeLeft = 900000 - (new Date().getTime() - localStorage.getItem('startTime'))
+      const timeLeft = 300000 - (new Date().getTime() - localStorage.getItem('startTime'))
       if (timeLeft > 0) { //if so, disable buttons and start time out w remaining time
         this.setState({disabled: true});
         setTimeout(() => {
@@ -82,7 +88,7 @@ class Master extends React.Component {
     setTimeout(() => {
       this.setState({disabled: false})
       localStorage.removeItem('startTime')
-    }, 900000);
+    }, 300000);
   }
 
   handleChange(event) { //react re-renders after every change
@@ -91,28 +97,29 @@ class Master extends React.Component {
   }
 
   handleSubmit(event) {
-    const unsubscribe = dbRef.add({
+    dbRef.add({
       text: this.state.value,
       timestamp: new Date().getTime() //Unix timestamp
     });
     event.preventDefault(); //prevents submit from navigating to a different page
     this.setState({value: ''}); //clears value
     this.disableSubmit();
-    return () => unsubscribe();
   }
 
   handleDelete(id){
-    const unsubscribe = dbRef.doc(id).delete();
-    this.disableSubmit();
-    return () => unsubscribe();
+    //confirm alert
+    if (window.confirm('Are you sure you want to delete this word?')){
+      dbRef.doc(id).delete();
+      this.disableSubmit();
+    }
   }
 
   render() {
     return (
       <div>
         <div>
-          <p>Enter a single word to add to the story OR click on a word to remove it from the story. <br/><br/>
-          You can only perform one action every 15 minutes.</p>
+          <p>Enter a single word to add to the story or click on a word to remove it. <br/><br/>
+          Choose carefully––only one action can be performed every 5 minutes.</p>
         </div>
         <div>
           <UpdatedStory onClick={id => this.handleDelete(id)} disabled={this.state.disabled}/>
